Refactor main folder to components folder
