+++ app/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from 'main/App';
+import App from 'components/App';
 import * as serviceWorker from './serviceWorker';
 
 import { Drizzle, generateStore } from 'drizzle';
